refactor: enhance utility functions with improved and documentation
