+++ packages/cli/src/utils/index.ts
@@ -1,8 +1,8 @@
-export * from "./getItemTargetPath";
-export * from "./getPackageInfo";
-export * from "./getPackageManager";
-export * from "./getPathFile";
-export * from "./handleErrors";
-export * from "./installPackages";
-export * from "./logger";
-export * from "./runCommand";
+export { default as getFilePath } from './getFilePath';
+export { default as getPackageInfo } from './getPackageInfo';
+export { default as handleErrors } from './handleErrors';
+export { default as spinner } from './spinner';
+export { default as getPackageManager } from './getPackageManager';
+export { default as installPackages } from './installPackages';
+export * from './logger';
+export { default as runCommand } from './runCommand';
